backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
djangorestframework-simplejwt>=5.2,<6.0
python-dotenv>=1.0,<2.0
Pillow==11.3.0
